requirements.txt
# MCP and FastMCP dependencies
mcp
fastmcp
requests

# Python environment management
python-dotenv

# LangGraph and LangChain dependencies
langgraph
langchain-google-genai
langchain-core
langchain-mcp-adapters

# LlamaIndex dependencies
llama-index
llama-index-llms-google-genai
llama-index-tools-mcp

# Typing extensions
typing-extensions